+++ src/test/java/com/svetylkovo/rojo/matcher/beans/NestedInner.java
@@ -1,0 +1,39 @@
+package com.svetylkovo.rojo.matcher.beans;
+
+import com.svetylkovo.rojo.annotations.Group;
+import com.svetylkovo.rojo.annotations.Regex;
+
+@Regex("(\\d)(\\d)(\\d)")
+public class NestedInner {
+
+    @Group(1)
+    private int first;
+    @Group(2)
+    private int second;
+    @Group(3)
+    private int third;
+
+    public int getFirst() {
+        return first;
+    }
+
+    public void setFirst(int first) {
+        this.first = first;
+    }
+
+    public int getSecond() {
+        return second;
+    }
+
+    public void setSecond(int second) {
+        this.second = second;
+    }
+
+    public int getThird() {
+        return third;
+    }
+
+    public void setThird(int third) {
+        this.third = third;
+    }
+}